requirements.txt
asgiref<=3.6.0
asttokens<=2.2.1
charset-normalizer<=3.1.0
cryptography<=40.0.2
debugpy<=1.6.7
decorator<=5.1.1
executing<=1.2.0
gunicorn<=20.1.0
python-dateutil<=2.8.2
python-slugify<=8.0.1
requests<=2.29.0
sqlparse<=0.4.4
stone<=3.3.1
tzdata<=2023.3
urllib3<=1.26.15
uvicorn<=0.22.0
whitenoise<=6.5.0
Django<=4.2
django-environ==0.10.0
django-cors-headers<=4.1.0
djangorestframework<=3.14.0
python-decouple<=3.8
Pillow<=9.5.0


